+++ thold.php
@@ -41,9 +41,14 @@ include_once($config['base_path'] . '/plugins/thold/thold_functions.php');
 set_default_action();
 
 if (isset($_SERVER['HTTP_REFERER'])) {
-	if (preg_match('/(data_sources.php|graph_view.php|graph.php)/', $_SERVER['HTTP_REFERER'])) {
+	if (preg_match('/(data_sources.php|graph_view.php|thold_graph.php|graph.php)/', $_SERVER['HTTP_REFERER'])) {
+		cacti_log('The return location is: ' . $_SERVER['HTTP_REFERER']);
 		$_SESSION['data_return'] = $_SERVER['HTTP_REFERER'];
+	} else {
+		kill_session_var('data_return');
 	}
+} else {
+	kill_session_var('data_return');
 }
 
 if (isset_request_var('drp_action')) {
@@ -444,7 +449,14 @@ function do_actions() {
 			print "<tr><td colspan='2'><p><i>Operator Message:</i><br><textarea class='ui-state-default ui-corner-all' style='width:70%;height:50px;' area-multiline='true' rows='2' id='message' name='message'></textarea></p></td></tr>";
 		}
 
-		$save_html = "<input type='button' class='ui-button ui-corner-all ui-widget' value='" . __esc('Cancel', 'thold') . "' onClick='cactiReturnTo()'>";
+		if (isset($_SESSION['data_return'])) {
+			$returnTo = sanitize_uri($_SESSION['data_return']);
+		} else {
+			$returnTo = $config['url_path'] . 'plugins/thold/thold.php';
+		}
+
+		$save_html = "<input type='button' class='ui-button ui-corner-all ui-widget' value='" . __esc('Cancel', 'thold') . "' onClick='cactiReturnTo($returnTo)'>";
+
 		if (!empty($button)) {
 			$save_html .= "&nbsp;<input type='submit' class='ui-button ui-corner-all ui-widget' value='" . __esc('Continue', 'thold') . "' title='$button'>";
 		}